type fix for event

+++ contracts/smart-account/interfaces/modules/IAccountRecoveryModule.sol
@@ -109,7 +109,7 @@ interface IAccountRecoveryModule {
      */
     event SecurityDelayChanged(
         address indexed smartAccount,
-        uint48 securityDelay
+        uint24 securityDelay
     );
 
     /**
